# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Aggregator API",
        "description": "REST API for a Laravel-based news aggregator service.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/preferences": {
            "get": {
                "tags": [
                    "Preferences"
                ],
                "summary": "Save user news preferences",
                "description": "Stores the user's preferred category and news source for personalized feeds.",
                "operationId": "94ae0966450c96135b23274e389b9ce4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category",
                                    "source"
                                ],
                                "properties": {
                                    "category": {
                                        "type": "string",
                                        "example": "technology"
                                    },
                                    "source": {
                                        "type": "string",
                                        "example": "News API"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Preference successfully saved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "category": {
                                                    "type": "string",
                                                    "example": "technology"
                                                },
                                                "source": {
                                                    "type": "string",
                                                    "example": "News API"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/preferences": {
            "post": {
                "tags": [
                    "Preferences"
                ],
                "summary": "Save user preferences",
                "description": "Store a newly created resource in storage.",
                "operationId": "d03dcea4a5a3eefea8e4407de1a92d2c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category": {
                                        "type": "string",
                                        "example": "technology"
                                    },
                                    "source": {
                                        "type": "string",
                                        "example": "News API"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Preference stored",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "category": {
                                                    "type": "string"
                                                },
                                                "source": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
...
}